projectManager: Add basic functionality
Add a projectManager module to handle a workspace containing
multiple projects. The user can enable/disable projects, edit
the project content, add or remove files while the manager tracks
the changes in the project and stores the related hdl source files
and constraint files in a separate project info file.